requirements.txt
# Intelligent Web Scraper - Requirements
# Python 3.13 required
# Last updated: February 2026

# Web Scraping
beautifulsoup4>=4.12.0,<5.0.0
requests>=2.31.0,<3.0.0
lxml>=5.0.0,<6.0.0

# Environment Management
python-dotenv>=1.0.0,<2.0.0

# NVIDIA API Integration
# No additional packages needed - using requests for direct API calls
# Get your free NVIDIA API keys from: https://build.nvidia.com/
